+++ tests/ui/test_deck_import_wizard.py
@@ -13,6 +13,7 @@
 # You should have received a copy of the GNU General Public License
 # along with this program. If not, see <http://www.gnu.org/licenses/>.
 
+import collections
 import itertools
 import typing
 from unittest.mock import MagicMock
@@ -50,12 +51,13 @@ def create_and_show_wizard(qtbot: QtBot, card_db: CardDatabase, cards: StringLis
 def test_going_back_to_textual_deck_list_resets_parsed_cards_model(qtbot: QtBot, card_db: CardDatabase):
     wizard = create_and_show_wizard(qtbot, card_db, ["regular_english_card"])
     deck_list = "1 Fury Sliver"
-    _select_magic_online_parser(qtbot, wizard)
-    _move_wizard_forward(qtbot, wizard)
     _input_deck_list(qtbot, wizard, deck_list)
+    _move_wizard_forward(qtbot, wizard)
+    _select_magic_online_parser(qtbot, wizard)
     _move_wizard_forward(qtbot, wizard)
     list_model = wizard.summary_page.card_list
     _validate_model_content(list_model)
+    list_model.clear()
     _move_wizard_backward(qtbot, wizard)
     _move_wizard_forward(qtbot, wizard)
     assert_that(wizard.summary_page.card_list, is_(same_instance(list_model)))
@@ -87,8 +89,8 @@ def _select_magic_online_parser(qtbot: QtBot, wizard: DeckImportWizard):
 
 def _select_magic_arena_parser(qtbot: QtBot, wizard: DeckImportWizard):
     page = wizard.select_deck_parser_page
-    with qtbot.wait_signal(page.completeChanged, timeout=100):
-        cb: QCheckBox = page.select_parser_mtg_arena
+    cb: QCheckBox = page.select_parser_mtg_arena
+    with qtbot.wait_signal(cb.clicked, timeout=100):
         cb.click()
     assert_that(page.select_parser_mtg_arena.isChecked())
     assert_that(page.complete, is_(True))
@@ -154,7 +156,7 @@ def _validate_model_content(list_model):
 class DeckReceiver(QObject):
     def __init__(self, parent: QObject = None):
         super(DeckReceiver, self).__init__(parent)
-        self.deck: typing.Counter[Card] = None
+        self.deck: typing.Counter[Card] = collections.Counter()
 
     def on_deck_received(self, deck: typing.Counter[Card]):
         self.deck = deck
@@ -163,9 +165,9 @@ class DeckReceiver(QObject):
 def test_selecting_different_printing_works(qtbot: QtBot, card_db: CardDatabase):
     wizard = create_and_show_wizard(qtbot, card_db, ["regular_english_card", "regular_english_card_reprint"])
     deck_list = "2 Fury Sliver (TSP) 157"
-    _select_magic_arena_parser(qtbot, wizard)
-    _move_wizard_forward(qtbot, wizard)
     _input_deck_list(qtbot, wizard, deck_list)
+    _move_wizard_forward(qtbot, wizard)
+    _select_magic_arena_parser(qtbot, wizard)
     _move_wizard_forward(qtbot, wizard)
     table_view: QTableView = wizard.summary_page.parsed_cards_table
     cell_position = QPoint(
@@ -233,9 +235,9 @@ def test_complete_button_disabled_if_zero_cards_identified(qtbot: QtBot, card_db
     """
     wizard = create_and_show_wizard(qtbot, card_db, ["regular_english_card"])
     deck_list = "Invalid deck list"
-    _select_magic_arena_parser(qtbot, wizard)
-    _move_wizard_forward(qtbot, wizard)
     _input_deck_list(qtbot, wizard, deck_list)
+    _move_wizard_forward(qtbot, wizard)
+    _select_magic_arena_parser(qtbot, wizard)
     _move_wizard_forward(qtbot, wizard)
     table_view: QTableView = wizard.summary_page.parsed_cards_table
     assert_that(table_view.model().rowCount(), is_(0), "Setup failed: Parsed deck model must be empty!")
@@ -249,9 +251,9 @@ def test_complete_button_enabled_if_one_card_identified(qtbot: QtBot, card_db: C
     """
     wizard = create_and_show_wizard(qtbot, card_db, ["regular_english_card"])
     deck_list = "1 Fury Sliver (TSP) 157"
-    _select_magic_arena_parser(qtbot, wizard)
-    _move_wizard_forward(qtbot, wizard)
     _input_deck_list(qtbot, wizard, deck_list)
+    _move_wizard_forward(qtbot, wizard)
+    _select_magic_arena_parser(qtbot, wizard)
     _move_wizard_forward(qtbot, wizard)
     table_view: QTableView = wizard.summary_page.parsed_cards_table
     assert_that(table_view.model().rowCount(), is_(1), "Setup failed: Parsed deck model must not be empty!")
@@ -267,9 +269,9 @@ def test_complete_state_updates_when_deck_list_updated_to_contain_cards(qtbot: Q
     wizard = create_and_show_wizard(qtbot, card_db, ["regular_english_card"])
     invalid_deck_list = "Invalid deck list"
     valid_deck_list = "1 Fury Sliver (TSP) 157"
-    _select_magic_arena_parser(qtbot, wizard)
-    _move_wizard_forward(qtbot, wizard)
     _input_deck_list(qtbot, wizard, invalid_deck_list)
+    _move_wizard_forward(qtbot, wizard)
+    _select_magic_arena_parser(qtbot, wizard)
     _move_wizard_forward(qtbot, wizard)
     table_view: QTableView = wizard.summary_page.parsed_cards_table
     assert_that(table_view.model().rowCount(), is_(0), "Setup failed: Parsed deck model must be empty!")
@@ -278,7 +280,10 @@ def test_complete_state_updates_when_deck_list_updated_to_contain_cards(qtbot: Q
 
     # Transition from invalid to valid state
     _move_wizard_backward(qtbot, wizard)
+    _move_wizard_backward(qtbot, wizard)
     _input_deck_list(qtbot, wizard, valid_deck_list)
+    _move_wizard_forward(qtbot, wizard)
+    _select_magic_arena_parser(qtbot, wizard)
     _move_wizard_forward(qtbot, wizard)
     assert_that(table_view.model().rowCount(), is_(1), "Setup failed: Parsed deck model must not be empty!")
     assert_that(wizard.summary_page.isComplete(), is_(True))
@@ -286,7 +291,10 @@ def test_complete_state_updates_when_deck_list_updated_to_contain_cards(qtbot: Q
 
     # Transition from valid to invalid state
     _move_wizard_backward(qtbot, wizard)
+    _move_wizard_backward(qtbot, wizard)
     _input_deck_list(qtbot, wizard, invalid_deck_list)
+    _move_wizard_forward(qtbot, wizard)
+    _select_magic_arena_parser(qtbot, wizard)
     _move_wizard_forward(qtbot, wizard)
     assert_that(table_view.model().rowCount(), is_(0), "Setup failed: Parsed deck model must be empty!")
     assert_that(wizard.summary_page.isComplete(), is_(False))
@@ -299,9 +307,9 @@ def test_custom_re_parser_works(qtbot: QtBot, card_db: CardDatabase):
     wizard = create_and_show_wizard(qtbot, card_db, ["regular_english_card", "regular_english_card_reprint"])
     cards = card_db.get_cards_from_data(CardIdentificationData("en", "Fury Sliver"))
     wizard.select_deck_parser_page.image_db.filter_already_downloaded.return_value = cards
-    _select_generic_re_parser(qtbot, wizard, valid_re, True)
-    _move_wizard_forward(qtbot, wizard)
     _input_deck_list(qtbot, wizard, deck_list, enable_print_guessing=True)
+    _move_wizard_forward(qtbot, wizard)
+    _select_generic_re_parser(qtbot, wizard, valid_re, True)
     _move_wizard_forward(qtbot, wizard)
     table_view: QTableView = wizard.summary_page.parsed_cards_table
     assert_that(table_view.model().rowCount(), is_(1), "Setup failed: Parsed deck model must not be empty!")
@@ -327,6 +335,9 @@ def generate_test_cases_for_test_custom_re_parser_accepts_valid_re():
 @pytest.mark.parametrize("valid_re", generate_test_cases_for_test_custom_re_parser_accepts_valid_re())
 def test_custom_re_parser_accepts_valid_re(qtbot, card_db, valid_re: str):
     wizard = create_and_show_wizard(qtbot, card_db, ["regular_english_card", "regular_english_card_reprint"])
+    deck_list = "Fury Sliver"
+    _input_deck_list(qtbot, wizard, deck_list, enable_print_guessing=True)
+    _move_wizard_forward(qtbot, wizard)
     _select_generic_re_parser(qtbot, wizard, valid_re, True)
 
 
@@ -340,4 +351,7 @@ def test_custom_re_parser_accepts_valid_re(qtbot, card_db, valid_re: str):
 ])
 def test_custom_re_parser_declines_non_identifying_re(qtbot: QtBot, card_db: CardDatabase, invalid_re: str):
     wizard = create_and_show_wizard(qtbot, card_db, ["regular_english_card", "regular_english_card_reprint"])
+    deck_list = "Fury Sliver"
+    _input_deck_list(qtbot, wizard, deck_list, enable_print_guessing=True)
+    _move_wizard_forward(qtbot, wizard)
     _select_generic_re_parser(qtbot, wizard, invalid_re, False)