+++ doc/changelog.md
@@ -1,5 +1,20 @@
 # Changelog
 
+# Next version (in development)
+
+## New features
+
+- Automatic deck list downloads. The deck list import wizard now has an input field that accepts
+  links to deck lists on various deck list database websites. 
+  - Currently supported are [Scryfall](https://scryfall.com),
+    [MTGGoldfish](https://www.mtggoldfish.com/) (`/deck` links only, `/archetype` is currently unsupported),
+    [mtg.wtf](https://mtg.wtf/), [TappedOut](https://tappedout.net/), [Moxfield](https://www.moxfield.com/) and
+    [Deckstats](https://deckstats.net/)
+
+## Changed features
+
+- Re-structured the deck list import wizard. The deck list input is now the first step.
+
 # Version 0.19.0 (2022-10-02)  <a name="v0_19_0"></a>
 
 ## New features