+++ doc/changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-# Next version (in development)
+# Version 0.20.0 (2022-10-09)  <a name="v0_20_0"></a>
 
 ## New features